move to util
