+++ resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/Config.scala
@@ -250,19 +250,6 @@ private[spark] object Config extends Logging {
       .stringConf
       .createOptional
 
-  // the definition of a label in DNS (RFC 1123).
-  private val dns1123LabelFmt = "[a-z0-9]([-a-z0-9]*[a-z0-9])?"
-
-  private val podConfValidator = (s"^$dns1123LabelFmt(\\.$dns1123LabelFmt)*$$").r.pattern
-
-  // The possible longest executor name would be "$prefix-exec-${Int.MaxValue}"
-  def isValidExecutorPodNamePrefix(prefix: String): Boolean = {
-    // 6 is length of '-exec-'
-    val reservedLen = Int.MaxValue.toString.length + 6
-    val validLength = prefix.length + reservedLen <= KUBERNETES_DNSNAME_MAX_LENGTH
-    validLength && podConfValidator.matcher(prefix).matches()
-  }
-
   val KUBERNETES_EXECUTOR_POD_NAME_PREFIX =
     ConfigBuilder("spark.kubernetes.executor.podNamePrefix")
       .doc("Prefix to use in front of the executor pod names. It must conform the rules defined " +
@@ -273,7 +260,7 @@ private[spark] object Config extends Logging {
         "so the length of the `podNamePrefix` needs to be <= 47(= 63 - 10 - 6).")
       .version("2.3.0")
       .stringConf
-      .checkValue(isValidExecutorPodNamePrefix,
+      .checkValue(KubernetesUtils.isValidExecutorPodNamePrefix,
         "must conform https://kubernetes.io/docs/concepts/overview/working-with-objects" +
           "/names/#dns-label-names and the value length <= 47")
       .createOptional
